% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baggr_plot.R
\name{add_color}
\alias{add_color}
\title{Add colors to baggr plots}
\usage{
add_color(p, add_color_to, style)
}
\arguments{
\item{p}{A ggplot object to add colors to}

\item{add_color_to}{A list of objects and colors to be added to the plots.
Lists are constructed such that objects and colors are given
in pairs. For example, a list to turn School A red and School B
blue would be constructed as \code{list('School A','red','School B','blue')}.
CSS hex numbers are also supported.}

\item{style}{The style of the plot. Currently, only plots of style \code{intervals} are
supported.}
}
\description{
Add colors to baggr plots
}
\examples{
bg <- baggr(schools)
colors <- list('School A','red','School B','blue')
plot(bg,add_color_to=colors)
}
