% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{add_or_multiply}
\alias{add_or_multiply}
\title{function to test codecov}
\usage{
add_or_multiply(num1, num2, type)
}
\arguments{
\item{num1}{one number to add or multiply}

\item{num2}{secondary numerb to add or multiply}

\item{type}{choose add or mulitply}
}
\value{
num1 + num2 or num1 * num2
}
\description{
function to test codecov
}
