% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baggr.R
\name{baggr}
\alias{baggr}
\title{Bayesian aggregate treatment effects model}
\usage{
baggr(data, model = NULL, prior = NULL, pooling = "partial",
  warnings = TRUE, outcome = "outcome", grouping = "site",
  treatment = "treatment", ...)
}
\arguments{
\item{data}{data frame with columns 'outcome', 'treatment' and 'site'}

\item{model}{choose from 'rubin', ...}

\item{prior}{WIP}

\item{pooling}{choose from `none`, `partial` (default) and `full`}

\item{warnings}{if `TRUE` (default), Stan warnings will be displayed}

\item{model}{choose from 'rubin', ...}
}
\value{
`stanfit` object with appropriate parameters
}
\description{
Estimate parameters in an ATE model
that's appropriate to the supplied
individual- or group-level data.
}
\details{
WIP
}
\author{
Witold Wiecek
}
