% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\docType{data}
\name{baggr_models}
\alias{baggr_models}
\title{Models currently supported by baggr
see \code{baggr()}}
\format{An object of class \code{character} of length 4.}
\usage{
baggr_models
}
\description{
Models currently supported by baggr
see \code{baggr()}
}
\keyword{datasets}
