% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baggr-package.R
\docType{package}
\name{baggr_package}
\alias{baggr_package}
\alias{baggr_package-package}
\title{baggr - a package for Bayesian meta-analysis}
\description{
This is \emph{baggr} (pronounced as \emph{bagger} or \emph{badger}), a Bayesian meta-analysis package for R using \href{https://mc-stan.org/}{Stan}. \emph{Baggr} is intended to be user-friendly and transparent so that it's easier to understand the models you are building and criticise them. The current version is a stable prototype of a tool that's in active development so we are counting on your feedback.
}
\details{
\emph{Baggr} provides a suite of Bayesian aggregation models for both summary statistics and full data sets to synthesise evidence collected from different groups, contexts or time periods.
}
\section{Getting help}{


This is only a sumple package help file.
For documentation of the main function for conducting analyses see \link{baggr}.
For description of models, data types and priors available in the package,
try the built-in vignette (\code{vignette("baggr")}).
}

\author{
Witold Wiecek, Rachael Meager
}
