% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{baggr_theme_set}
\alias{baggr_theme_set}
\alias{baggr_theme_get}
\alias{baggr_theme_update}
\alias{baggr_theme_replace}
\title{Set, get, and replace themes for baggr plots}
\usage{
baggr_theme_set(new = bayesplot::theme_default())

baggr_theme_get()

baggr_theme_update(...)

baggr_theme_replace(...)
}
\arguments{
\item{new}{New theme to use for all baggr plots}
}
\value{
The get method returns the current theme, but all of the
others invisibly return the old theme.
}
\description{
These functions get, set, and modify the ggplot2 themes
of the baggr plots. By default, these match the bayesplot's package
theme to make it easier to form cohesive graphs across this package
and others.
}
