% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble.R
\name{bubble}
\alias{bubble}
\title{Draw a bubble plot for a baggr object}
\usage{
bubble(bg, covariate, pred = TRUE, label = TRUE)
}
\arguments{
\item{bg}{A \link{baggr} object with at least one covariate}

\item{covariate}{Character: a name of one of the covariates present in \code{bg}}

\item{pred}{TRUE/FALSE; print the linear trend for the variable?}

\item{label}{A logical value that determines if group labels are added to the plot.}
}
\value{
A bubble plot (\code{ggplot})
}
\description{
Draw a bubble plot for a baggr object
}
\details{
The bubble plot shows the observed outcomes of individual
studies plotted against study-level covariates. The size of each point reflects
the precision of the study (inverse variance).
}
\examples{

schools$a <- rnorm(8)
bg <- baggr(schools, cov = "a")
bubble(bg, "a")

}
