% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{bubble_plot}
\alias{bubble_plot}
\title{Draw a bubble plot using ggplot syntax}
\usage{
bubble_plot(bg, covariates)
}
\arguments{
\item{bg}{A \link{baggr} object.}

\item{covariates}{Covariates for the \link{baggr} data.}
}
\value{
A bubble plot of 1/SE^2
}
\description{
Draw a bubble plot using ggplot syntax
}
\examples{

covariates <- rnorm(8)
bg <- baggr(schools)
bubble_plot(bg,covariates)

}
