% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microcredit.R
\docType{data}
\name{chicks}
\alias{chicks}
\title{Chickens: impact of electromagnetic field on calcium ion efflux in chicken brains}
\format{
An object of class \code{data.frame} with 38 rows and 7 columns.
}
\usage{
chicks
}
\description{
An experiment conducted by Blackman et al. (1988) and documented in the following
\href{https://github.com/VMatthijs/Slamming-the-sham}{GitHub repository} by Vakar
and Gelman. The dataset consists of a large number of experiments (\code{tau}, \code{se.tau})
repeated at varying wave frequencies. Sham experiments (\code{mu}, \code{se.mu}) are also
included, allowing us to compare performance of models with and without
control measurements.
}
\references{
Blackman, C. F., S. G. Benane, D. J. Elliott, D. E. House, and
M. M. Pollock.
“Influence of Electromagnetic Fields on the Efflux of Calcium Ions from Brain
Tissue in Vitro: A Three-Model Analysis Consistent with the Frequency
Response up to 510 Hz.” Bioelectromagnetics 9, no. 3 (1988): 215–27.
}
\keyword{datasets}
