% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_inputs.R
\name{convert_inputs}
\alias{convert_inputs}
\title{Convert inputs for baggr models}
\usage{
convert_inputs(data, model, quantiles, group = "group",
  outcome = "outcome", treatment = "treatment", test_data = NULL)
}
\arguments{
\item{data}{`data.frame`` with desired modelling input}

\item{model}{valid model name used by baggr;
see \code{?baggr} for allowed models}

\item{quantiles}{vector of quantiles to use (only applicable if \code{model = "quantiles"})}

\item{group}{name of the column with grouping variable}

\item{outcome}{name of column with outcome variable}

\item{treatment}{name of column with treatment variable}

\item{test_data}{same format as \code{data} argument, gets left aside for testing purposes (see \code{link{baggr}})}
}
\value{
data.frame of class baggr_data that baggr() uses
}
\description{
Convert inputs for baggr models
}
\details{
The conversions will typically happen automatically when data is fed to baggr()
function. This function can be used to explicitly obtain inputs formatted the way
that Stan models use data.
}
\author{
Witold Wiecek, Rachael Meager
}
