% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microcredit.R
\docType{data}
\name{data_spike}
\alias{data_spike}
\title{Spike & slab example dataset}
\format{
An object of class \code{data.frame} with 1500 rows and 4 columns.
}
\usage{
data_spike
}
\description{
Spike & slab example dataset
}
\keyword{datasets}
