% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_effects.R
\name{fixed_effects}
\alias{fixed_effects}
\title{Effects of covariates on outcome in baggr models}
\usage{
fixed_effects(bg, summary = FALSE, transform = NULL, interval = 0.95)
}
\arguments{
\item{bg}{a \code{\link[=baggr]{baggr()}} model}

\item{summary}{logical; if \code{TRUE} returns summary statistic instead of all MCMC samples}

\item{transform}{a transformation (R function) to apply to the result;
(this is commonly used when calling from other
plotting or printing functions)}

\item{interval}{uncertainty interval width (numeric between 0 and 1), if \code{summary=TRUE}}
}
\value{
A matrix: columns are covariate coefficients and rows are draws from the posterior distribution.
Number of rows depends on iterations in the MCMC (i.e. \code{x} in baggr(..., iter = x`)
}
\description{
Effects of covariates on outcome in baggr models
}
\seealso{
\link{treatment_effect} for overall treatment effect across groups,
\link{group_effects} for effects within each group,
\link{effect_draw} and \link{effect_plot} for predicted treatment effect in new group
(which you can condition on fixed effects using new data argument)
}
