% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel.R
\name{funnel}
\alias{funnel}
\title{Funnel plots for baggr models}
\usage{
funnel(bg, show = c("inputs", "posterior"), level = 0.95, label = FALSE)
}
\arguments{
\item{bg}{a \code{\link[=baggr]{baggr()}} model}

\item{show}{whether to plot raw study-level inputs (\code{"inputs"}) or
posterior summaries (\code{"posterior"})}

\item{level}{confidence level for reference lines}

\item{label}{logical: add study/group labels?}
}
\value{
A ggplot funnel plot for the supplied model
}
\description{
Funnel plots for baggr models
}
\examples{
bg <- baggr(schools, iter = 500, refresh = 0)
funnel(bg, label = TRUE)

}
