% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{get_n_samples}
\alias{get_n_samples}
\title{Extract number of samples from a baggr object}
\usage{
get_n_samples(x)
}
\arguments{
\item{x}{baggr fit to get samples from}
}
\description{
Extract number of samples from a baggr object
}
\details{
Checks for number of iterations and
number of Markov chains, returns maximum number
of valid samples
}
