% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baggr_compare.R
\name{get_order}
\alias{get_order}
\title{Separate out ordering so we can test directly}
\usage{
get_order(df_groups, hyper)
}
\arguments{
\item{df_groups}{data.frame of group effects used in \link{plot.baggr_compare}}

\item{hyper}{show parameter estimate? same as in \link{plot.baggr_compare}}
}
\description{
Separate out ordering so we can test directly
}
\details{
Given a set of effects measured by models, identifies the
model which has the biggest range of estimates and ranks groups
by those estimates, returning the order
}
