% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{get_y}
\alias{get_y}
\title{Get Y from various models}
\usage{
get_y(x)
}
\arguments{
\item{x}{model to get y}
}
\description{
Get Y from various models
}
\details{
grabs the relevant Y variable
for use with posterior or prior predictive checks.
}
