% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trt_effects.R
\name{hypermean}
\alias{hypermean}
\title{Return the hypermean of the treatments effects}
\usage{
hypermean(bg, transform = NULL, interval = 0.95, message = TRUE)
}
\arguments{
\item{bg}{A \link{baggr} model.}

\item{transform}{A transformation to apply to the result, should be an R function;
(this is commonly used when calling \code{hypermean} from other
plotting or printing functions)}

\item{interval}{The confidence inerval of the hypermean passed as any value between 0 and 1.}

\item{message}{A logical agrument that determines whether the result is printed or not.}
}
\value{
The hypermean of the \code{baggr} model with the interval specified by \code{interval}, with
a default interval of 95\%, as calculated by \link{treatment_effect}.
}
\description{
Return the hypermean of the treatments effects
}
