% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baggr_compare.R
\name{make_silent}
\alias{make_silent}
\title{Make an expression silent}
\usage{
make_silent(...)
}
\arguments{
\item{...}{expression to pass to make silent}
}
\description{
Make an expression silent
}
\details{
Runs the example while suppressing messages and warnings
while sinking the output to a temp file and then deleting
(this bit is handled by testthat)
}
\examples{
\dontrun {
tmp <- baggr(schools)
make_silent(tmp <- baggr(schools))
}
}
