% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microcredit.R
\docType{data}
\name{microcredit}
\alias{microcredit}
\title{7 studies on effect of microcredit supply}
\format{A data frame with 40267 rows, 7 study idenfitifiers and 7 outcomes}
\usage{
microcredit
}
\description{
Rachael will write this up.
}
\keyword{datasets}
