% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microcredit.R
\docType{data}
\name{microcredit_simplified}
\alias{microcredit_simplified}
\title{Simplified version of the microcredit dataset.}
\format{A data frame with 14224 rows, 7 study idenfitifiers and 1 outcome}
\usage{
microcredit_simplified
}
\description{
Includes only complete cases and only the consumer durables outcome variable.
}
\keyword{datasets}
