% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mint.R
\name{mint}
\alias{mint}
\title{Output mean, intervals and other summaries for matrix (by column) or vector}
\usage{
mint(y, int = 0.95, digits = NULL, median = FALSE, sd = FALSE)
}
\arguments{
\item{y}{matrix or a vector; for matrices, \code{mint} is done by-column}

\item{int}{probability interval (default is 95 percent) to calculate}

\item{digits}{number of significant digits to \link[=round]{base} values by.}

\item{median}{calculate median value?}

\item{sd}{calculate SD}
}
\description{
This function is just a convenient shorthand for getting typical statistics.
}
\examples{
mint(rnorm(100, 12, 5))

}
