% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trt_effects.R
\name{mutau_cor}
\alias{mutau_cor}
\title{Correlation between mu and tau in a baggr model}
\usage{
mutau_cor(bg, summary = FALSE, interval = 0.95)
}
\arguments{
\item{bg}{a \code{\link[=baggr]{baggr()}} model where \code{model = "mutau"}}

\item{summary}{logical; if TRUE returns summary statistics as explained below.}

\item{interval}{uncertainty interval width (numeric between 0 and 1), if summarising}
}
\value{
a vector of values
}
\description{
Correlation between mu and tau in a baggr model
}
