% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv.R
\name{plot.baggr_cv}
\alias{plot.baggr_cv}
\title{Plot LOO CV results for baggr models}
\usage{
\method{plot}{baggr_cv}(x, y, ...)
}
\arguments{
\item{x}{output from \link{loocv} that has \code{return_models = TRUE}}

\item{y}{ignore}

\item{...}{}
}
\value{

}
\description{
Plot LOO CV results for baggr models
}
