% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_calculations.R
\name{power_calc_example}
\alias{power_calc_example}
\title{IDEA - NOT FOR USE -}
\usage{
power_calc_example()
}
\description{
Should baggr include power calculations?
Danny Toomey, April 2023
}
\details{
One of the main reasons researchers choose to perform meta-analyses is the potential
to detect effects individual studies cannot by increaseing statistical power. While
this is true of fixed-effects models, it is not necessarily true of random effects
models, which are far more common.

There are multiple ways to evaluate the possible effects of insufficient sample size
or excessive heterogeniety in a random effects model, which are the key variables that
would effect its statistical power. An argument against including a power measure would
be that heterogeniety is already given in baggr models and evaluating datasets with low
sample sizes is one of the key benefits of Bayesian analyses. It is possible that a low
statistical power is why a user would be interesting in building a Bayesian model in the
first place.

However, it is also possible that a user finds a power analysis informative in explaining
unexpected results. Consider a user is expecting to find significant effects in a Bayesian
model of Yusuf et al 1985, and they are surprised when they do not. However, a power
analysis shows the meta-analysis has a power of 0.26. The user may then infer that although
a Bayesian model may be less sensitive to small sample size, there may still be significant
limitations in its representation of findings in the field.

The \link{power_calc_example} function illustrates an example of this use case and proposes
an implementation for using power analyses to contextualize baggr models.

In general, a power metric may not be useful to all models. However, in cases where users observe
an unexpected effect, the power of the meta-analysis may provide useful context.

References:
\enumerate{
\item Quintana, D. How to calculate statistical power for your meta-analysis. Medium. 2017 Jul.
Available online: https://towardsdatascience.com/how-to-calculate-statistical-power-for-your-meta-analysis-e108ee586ae8
\item Jackson D, Turner R. Power analysis for random‐effects meta‐analysis.
Research synthesis methods. 2017 Sep;8(3):290-302. DOI: 10.1002/jrsm.1240.
}
}
