% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_mutau}
\alias{predict_mutau}
\title{Predict function for the mu & tau model}
\usage{
predict_mutau(x, nsamples, newdata = NULL, allow_new_levels = TRUE)
}
\arguments{
\item{x}{model to predict from}

\item{nsamples}{number of samples to predict}

\item{newdata}{new data to predict, defaults to NULL}

\item{allow_new_levels}{allow the predictive of new, unobserved groups}
}
\description{
Predict function for the mu & tau model
}
