% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_quantiles}
\alias{predict_quantiles}
\title{Predict function for the quantiles model}
\usage{
predict_quantiles(x, newdata = NULL, allow_new_levels = T, nsamples,
  ...)
}
\arguments{
\item{x}{model to predict from}

\item{newdata}{new data to predict, defaults to NULL}

\item{allow_new_levels}{allow the predictive of new, unobserved groups}

\item{nsamples}{number of samples to predict}

\item{...}{additional arguments, unused for now}
}
\description{
Predict function for the quantiles model
}
