% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_rubin}
\alias{predict_rubin}
\title{Predict function for the rubin model}
\usage{
predict_rubin(x, newdata = NULL, allow_new_levels = T, nsamples, ...)
}
\arguments{
\item{x}{model to predict from}

\item{newdata}{new data to predict, defaults to NULL}

\item{allow_new_levels}{allow the predictive of new, unobserved groups}

\item{nsamples}{number of samples to predict}
}
\description{
Predict function for the rubin model
}
