% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_unknown}
\alias{predict_unknown}
\title{Predict method for model that is unknown or not implemented}
\usage{
predict_unknown(x)
}
\arguments{
\item{x}{baggr model to generate predictions from}
}
\description{
Predict method for model that is unknown or not implemented
}
