% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baggr_compare.R
\name{print.baggr_compare}
\alias{print.baggr_compare}
\title{Print method for baggr_compare models}
\usage{
\method{print}{baggr_compare}(x, digits, ...)
}
\arguments{
\item{x}{baggr_compare model}

\item{digits}{number of significant digits for effect estimates}

\item{...}{other parameters passed to print}
}
\description{
Print method for baggr_compare models
}
