% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv.R
\name{print.baggr_cv}
\alias{print.baggr_cv}
\title{Print baggr cv objects nicely}
\usage{
\method{print}{baggr_cv}(x, digits = 3, ...)
}
\arguments{
\item{x}{\code{baggr_cv} object obtained from \link{loocv} to print}

\item{digits}{number of digits to print}

\item{...}{Unused, ignore}
}
\description{
Print baggr cv objects nicely
}
