% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv.R
\name{print.compare_baggr_cv}
\alias{print.compare_baggr_cv}
\title{Print baggr_cv comparisons}
\usage{
\method{print}{compare_baggr_cv}(x, digits = 3, ...)
}
\arguments{
\item{x}{baggr_cv comparison to print}

\item{digits}{number of digits to print}
}
\description{
Print baggr_cv comparisons
}
