% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baggr_compare.R
\name{print.plot_list}
\alias{print.plot_list}
\title{Print list of baggr plots}
\usage{
\method{print}{plot_list}(x)
}
\arguments{
\item{x}{list of plots to print}

\item{...}{ignored for now}
}
\description{
Print list of baggr plots
}
\details{
prints plots in a loop, internal use only
}
