% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_dist.R
\name{print_dist}
\alias{print_dist}
\title{Output a distribution as a string}
\usage{
print_dist(dist)
}
\arguments{
\item{dist}{distribution name, one of \link{priors}}
}
\value{
Character string like \code{normal(0, 10^2)}.
}
\description{
Used for printing nicely formatted outputs when reporting results etc.
}
