% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_dist.R
\name{priors}
\alias{priors}
\alias{multinormal}
\alias{lkj}
\alias{normal}
\alias{cauchy}
\alias{uniform}
\title{Prior distributions in baggr}
\usage{
multinormal(location, Sigma)

lkj(shape, order = NULL)

normal(location, scale)

cauchy(location, scale)

uniform(lower, upper)
}
\arguments{
\item{location}{Mean for normal, multivariate normal (in which case \code{location} is a vector)
and for Cauchy distributions}

\item{Sigma}{Variance-covariance matrix for multivariate normal.}

\item{shape}{Shape parameter for LKJ}

\item{order}{Order of LKJ matrix (typically it does not need to be specified,
as it is inferred directly in the model)}

\item{scale}{SD for Normal, scale for Cauchy}

\item{lower}{Lower bound for Uniform}

\item{upper}{Upper bound for Uniform}
}
\description{
This page provides a list of all available distributions
that can be used to specify priors in \code{\link[=baggr]{baggr()}}. These convenience functions
are designed only to let user write the priors in the most "natural" way. Apart from
passing on the arguments, their only other role is rudimentary check if the distribution
is specified correctly.
}
