% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baggr_code.R
\name{show_model}
\alias{show_model}
\title{Show Stan code for baggr models or objects}
\usage{
show_model(model)
}
\arguments{
\item{model}{either a \code{baggr} object (fitted model) or one of
\code{"rubin"}, \code{"mutau"}, \code{"individual"}}
}
\value{
Nothing is returned in R. Stan code will be opened externally
(e.g. via notepad).
}
\description{
Show Stan code for baggr models or objects
}
