% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_effects.R
\name{study_effects}
\alias{study_effects}
\title{Extract baggr study effects}
\usage{
study_effects(bg, summary = FALSE, interval = 0.95)
}
\description{
Internal function used as a helper for plotting and printing of results.
}
