% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_ma.R
\name{summarise_ma}
\alias{summarise_ma}
\title{Convert from individual to summary data in meta-analyses}
\usage{
summarise_ma(data)
}
\arguments{
\item{data}{data.frame of individual-level observations
with columns \code{outcome} (numeric),
\code{treatment} (values 0 and 1) and \code{site}}
}
\value{
data.frame with columns \code{mu}, \code{se.mu}, \code{tau} and \code{se.tau}
}
\description{
Allows only one-way conversion from full to summary data.
Input must be pre-formatted appropriately.
}
\details{
The conversions will typically happen automatically when data is fed to baggr()
function. This function can be used to explicitly convert from full to reduced
data without analysing it in any model.
}
\seealso{
\code{\link{convert_inputs}}
}
\author{
Witold Wiecek, Rachael Meager
}
