% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_quantiles_data.R
\name{summarise_quantiles_data}
\alias{summarise_quantiles_data}
\title{summarise_quantiles_data
Given individual level data, return list of summary statistics
Of quantile means and Sigma's, as well as K, N
Already formatted as a valid input into a Stan model of quantiles}
\usage{
summarise_quantiles_data(data, quantiles, grouping = "site",
  outcome = "outcome", treatment = "treatment")
}
\description{
summarise_quantiles_data
Given individual level data, return list of summary statistics
Of quantile means and Sigma's, as well as K, N
Already formatted as a valid input into a Stan model of quantiles
}
