% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Test function to see how adding functions works}
\usage{
test(x, y)
}
\arguments{
\item{x}{changing description to test roxygen}

\item{y}{another number}
}
\value{
Adds x and y
}
\description{
Test function to see how adding functions works
}
