% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microcredit.R
\docType{data}
\name{yusuf}
\alias{yusuf}
\title{Yusuf et al: beta-blockers and heart attacks}
\format{
An object of class \code{data.frame} with 21 rows and 5 columns.
}
\usage{
yusuf
}
\description{
This replicates Table 6 from the famous Yusuf et al. (1985), removing one trial (Snow)
that had NA observations only. The paper is notable for application of rare-event
corrections, which we discuss more in package vignette \code{baggr_binary}.
}
\references{
Yusuf, S., Peto, R., Lewis, J., Collins, R., & Sleight, P. (1985).
Beta blockade during and after myocardial infarction:
An overview of the randomized trials.
Progress in Cardiovascular Disease, 27(5), 335–371.
}
\keyword{datasets}
