% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baggr-package.R
\docType{package}
\name{baggr-package}
\alias{baggr-package}
\alias{baggr}
\title{The 'baggr' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. http://mc-stan.org
}
