% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baggr.R
\name{baggr}
\alias{baggr}
\title{Bayesian aggregate treatment effects model}
\usage{
baggr(data, model = NULL, prior = NULL, pooling = "partial",
  log = FALSE, cfb = FALSE, joint_prior = TRUE,
  standardise = FALSE, test_data = NULL, quantiles = seq(0.05, 0.95,
  0.1), outcome = "outcome", group = "group",
  treatment = "treatment", baseline = NULL, warn = TRUE, ...)
}
\arguments{
\item{data}{data frame with summary or individual level data to meta-analyse}

\item{model}{if \code{NULL}, detected automatically from input data
otherwise choose from \code{"rubin"}, \code{"mutau"}, \code{"individual"}, \code{"quantiles"}}

\item{prior}{list of prior arguments passed directly to each model as a vector of strings (see Details)}

\item{pooling}{choose from \code{none}, \code{partial} (default) and \code{full}}

\item{joint_prior}{If \code{TRUE}, \code{mu} and \code{tau} will have joint distribution.
If \code{FALSE}, they have independent priors. Ignored if no control
(\code{mu}) data exists.}

\item{test_data}{data for cross-validation; NULL for no validation, otherwise a data frame
with the same columns as \code{data} argument (see \code{\link[baggr]{loocv}} for automation)}

\item{quantiles}{if \code{model = "quantiles"}, a vector indicating which quantiles of data to use
(with values between 0 and 1)}

\item{outcome}{character; column name in (individual-level) \code{data} with outcome variable values}

\item{group}{character; column name in \code{data} with grouping factor;
it's necessary for individual-level data, for summarised data
it will be used as labels for groups when displaying results}

\item{treatment}{character; column name in (individual-level) \code{data} with treatment factor;}

\item{warn}{print warning if Rhat exceeds 1.05}

\item{...}{extra options passed to Stan function, e.g. \code{control = list(adapt_delta = 0.99)},
number of iterations etc.}
}
\value{
\code{baggr} class structure: list with Stan model fit embedded inside it,
alongside input data, pooling metrics, various model properties
}
\description{
Estimate parameters of an ATE model
that's appropriate to the supplied
individual- or group-level data.
(For overall package help file see \code{?baggr_package})
}
\details{
Most of data preparation steps can be done automatically through \code{\link[baggr]{prepare_ma}}.
Same function can also be used to convert individual level data to aggregate data.
While the preparation step is optional it will also automatically format data inputs to be
immediately recognisable by \code{baggr()}.

The "prior" argument currently requires you to specify all priors if you are passing any manually. They should be specified as a vector of string arguments with names specific to the model you are using (see the vignette for more details).

(OTHER SECTIONS OF THIS HELP ARE STILL WORK IN PROGRESS)
}
\examples{
df_pooled <- data.frame("tau" = c(1, -1, .5, -.5, .7, -.7, 1.3, -1.3),
"se" = rep(1, 8),
"state" = datasets::state.name[1:8])
baggr(df_pooled) #automatically detects the input data
# correct labels & passing some options to Stan
baggr(df_pooled, group = "state", iter = 200)

}
\author{
Witold Wiecek, Rachael Meager
}
