% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baggr_compare.R
\name{baggr_compare}
\alias{baggr_compare}
\title{Compare a(ny) number of baggr models side by side}
\usage{
baggr_compare(..., compare = "site", style = "areas", arrange = "single")
}
\arguments{
\item{...}{Either any number of objects of class `baggr`
(you can name your objects, see example below)
or the same arguments you'd pass to baggr()
function, but with `pooling = ...` omitted.
In the latter case 3 models will be run, with
pooling set to `none`, `partial` and `full`.}

\item{compare}{can be 'site' (compare different sites in each model),
or 'ate' (compare aggregate treatment effects only)}

\item{style}{What kind of plot to display - see options for `plot.baggr`}

\item{arrange}{If `single`, generate a single plot, if `grid`
display multiple plots side-by-side.}
}
\value{
ggplot is plotted and an extra comparison printed.
        Returned value is a list that contains
        the models, the plot and printed summaries. (WIP)
}
\description{
Compare a(ny) number of baggr models side by side
}
\author{
Witold Wiecek
}
