% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baggr-package.R
\docType{package}
\name{baggr_package}
\alias{baggr_package}
\alias{baggr_package-package}
\title{baggr - a package for meta-analysis}
\description{
(Work in progress.) This is overall package help file.
For documentation of \code{baggr()} function for conducting
analyses see \code{?baggr}
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. http://mc-stan.org
}
\author{
Witold Wiecek, Rachael Meager
}
