% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble.R
\name{bubble}
\alias{bubble}
\title{Bubble plots for meta-regression models}
\usage{
bubble(bg, covariate, fit = TRUE, label = TRUE)
}
\arguments{
\item{bg}{a \code{\link[=baggr]{baggr()}} model using summary-level data, with covariates}

\item{covariate}{one of the covariates present in the model}

\item{fit}{logical: show mean model prediction? (slope is mean estimate of \code{\link[=fixed_effects]{fixed_effects()}}, intercept is \code{\link[=hypermean]{hypermean()}});
if you have more than two groups and the covariate is a factor, this will be ignored}

\item{label}{logical: label study/group names?}
}
\value{
A simple bubble plot in \code{ggplot} style.
Dot sizes are proportional to inverse of variance of each study (more precise studies are larger).
}
\description{
Bubble plots for meta-regression models
}
\seealso{
\code{\link[=labbe]{labbe()}} for an exploratory plot of binary data in similar style
}
