% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{bubble_plot}
\alias{bubble_plot}
\title{Draw a bubble plot using ggplot syntax}
\usage{
bubble_plot(bg, covariates, regression = TRUE, interval = 0.95)
}
\arguments{
\item{bg}{A \link{baggr} object.}

\item{covariates}{Covariates for the \link{baggr} data.}

\item{regression}{A logical value that sets an overlying linear regression and
confidence interval on or off.}

\item{intervals}{The confidence interval used in the linear regression,
given as a value between 0 and 1.}
}
\value{
A bubble plot of 1/SE^2 for Tau as a function of the given covariates with
linear regression and confidence interval.
}
\description{
Draw a bubble plot using ggplot syntax
}
\examples{

covariates <- rnorm(8)
bg <- baggr(schools)
bubble_plot(bg,covariates)

}
