% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_inputs.R
\name{convert_inputs}
\alias{convert_inputs}
\title{Convert inputs for baggr Stan models}
\usage{
convert_inputs(data, model, grouping = "site", outcome = "outcome",
  treatment = "treatment", standardise = FALSE)
}
\arguments{
\item{data}{data.frame with desired modelling input}

\item{model}{valid model name used by baggr;
see \code{?baggr} for allowed models}

\item{grouping}{name of the column with grouping variable}

\item{outcome}{name of column with outcome variable}

\item{treatment}{name of column with treatment variable}

\item{standardise}{logical; whether to standardise data when converting}
}
\value{
data.frame of class baggr_data that baggr() uses
}
\description{
Allows conversions from full to reduced (summary) data and
in summary between long (control and treatment on separate rows)
and wide data (control and treatment in separate columns).
}
\details{
The conversions will typically happen automatically when data is fed to baggr()
function. This function can be used to explicitly convert from full to reduced
data without analysing it in a model.
}
\author{
Witold Wiecek
}
