% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{effect_draw}
\alias{effect_draw}
\title{Make posterior draws for treatment effect}
\usage{
effect_draw(x)
}
\arguments{
\item{x}{A \code{baggr} class object.}
}
\description{
Make posterior draws for treatment effect
}
