% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trt_effects.R
\name{effect_draw}
\alias{effect_draw}
\title{Make posterior draws for treatment effect}
\usage{
effect_draw(x, n)
}
\arguments{
\item{x}{A \code{baggr} class object.}

\item{n}{How many values to draw? The default is the same
as number of samples in the model (default is 2,000).}
}
\value{
A vector of possible values of the treatment effect.
}
\description{
This function takes the samples of hyperparameters of a \code{baggr} model
(commonly hypermean tau and hyper-SD sigma_tau) and simulates values of
new realisations of tau (a mean effect in some unobserved group).
}
