% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trt_effects.R
\name{hypermean}
\alias{hypermean}
\title{Return the hypermean of the treatments effects}
\usage{
hypermean(bg, transform = NULL, interval = 0.95, message = FALSE)
}
\arguments{
\item{bg}{A \link{baggr} model.}

\item{interval}{The confidence inerval of the hypermean passed as any value between 0 and 1.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     The hypermean of the `baggr` model with the interval specified 
             by `interval`, with a default interval of 95\%.
}\if{html}{\out{</div>}}
}
\description{
Return the hypermean of the treatments effects
}
