% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv.R
\name{loo_compare}
\alias{loo_compare}
\title{Compare fitted models on loo}
\usage{
loo_compare(x, ...)
}
\arguments{
\item{x}{An object of class "baggr_cv" or a list of such objects.}

\item{...}{Additional objects of class "baggr_cv"}
}
\description{
Compare fitted models on loo
}
