% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv.R
\name{loocv}
\alias{loocv}
\title{Leave one out cross-validation for \code{baggr} models}
\usage{
loocv(data, return_models = FALSE, ...)
}
\arguments{
\item{data}{Input data frame - same as for \link{baggr} function.}

\item{return_models}{logical; if FALSE, summary statistics will be returned and the
models discarded;
if TRUE, a list of models will be returned alongside summaries}

\item{...}{Additional arguments passed to \link{baggr}.}
}
\value{
log predictive density value, an object of class \code{baggr_cv};
full model, prior values and \emph{lpd} of each model are also returned
these can be examined by using \code{attributes()} function
}
\description{
Performs leave-one-out cross-validation on a \code{baggr} model at the group level.
This function automatically runs K \code{baggr} models, leaving out one group at a time,
and then calculating log predictive density for that group (see Gelman \emph{et al} 2014).
The main output is -2 times the log predictive density averaged over the K models, which corresponds to the Watanabe-Aikake Information Criterion.
This function takes in the same arguments as \code{baggr()}, plus an option (\code{return_models}) for whether to return all the models or just the summary statistics.
}
\details{
This function can be used to understand how any one group affects the overall result, as well as how well the model predicts
the omitted group. Because this function runs K models in total, it is recommended to set \code{mc.cores} option before running \code{loocv}, e.g. \code{options(mc.cores = 4)}.
Even with this option enabled, this function often has a long runtime even for simple examples.
The main output is -2 times the log predictive density averaged over K models, which corresponds to the Watanabe-Aikake Information Criterion.
A WAIC value closer to zero (i.e. a smaller number in magnitude) means a better fit.

More data are stored in \code{loocv()} output, and can be accessed via \code{attributes()}, e.g. the mean treatment effects, their variability and \emph{lpd} for each model that are stored in the attribute \code{df}.
}
\examples{
cv <- loocv(schools, return_models = FALSE, "rubin", pooling = "partial")
print(cv) #returns the lpd value
attributes(cv) #more information is included in the object

}
\references{
Gelman A, Hwang J, Vehtari A.
Understanding predictive information criteria for Bayesian models.
Statistics and Computing. 2014 Nov 24(6):997-1016.
\href{http://www.stat.columbia.edu/~gelman/research/published/waic_understand3.pdf}{PDF link.}
}
\author{
Witold Wiecek
}
