% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microcredit.R
\docType{data}
\name{microcredit_simplified}
\alias{microcredit_simplified}
\title{Simplified version of the microcredit dataset.}
\format{A data frame with 14224 rows, 7 study identifiers and 1 outcome}
\usage{
microcredit_simplified
}
\description{
This dataframe contains the data used in Meager (2019) to estimate hierarchical
models on the data from 7 randomized controlled trials of expanding access to microcredit.
}
\details{
The columns include the group indicator which gives the name of the lead author on
each of the respective studies, the value of the household consumer durables
spending standardised to USD PPP in 2009 dollars per two weeks (these are flow variables),
and finally a treatment assignment status indicator.

The dataset has not otherwise been cleaned and therefore includes NAs and other
issues common to real data.

For more information on how and why these variables were chosen and standardised,
see Meager (2019) or consult the associated code repository:
\href{https://bitbucket.org/rmeager/aggregate-average-impacts-microcredit/src/master/}{link}

This dataset includes only complete cases and only the consumer durables outcome variable.
}
\references{
Meager, Rachael (2019) Understanding the average impact of microcredit expansions:
A Bayesian hierarchical analysis of seven randomized experiments. American Economic Journal:
Applied Economics, 11(1), 57-91.
}
\keyword{datasets}
