% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_baggr.R
\name{plot.baggr}
\alias{plot.baggr}
\title{Plotting method in baggr package}
\usage{
\method{plot}{baggr}(bg, style = "areas", ...)
}
\arguments{
\item{bg}{object of class \code{baggr}}

\item{style}{one of 'areas', 'intervals' or (WIP)}

\item{...}{extra arguments passed to bayesplot functions}
}
\value{
ggplot2 object
}
\description{
Plotting method in baggr package
}
