% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.baggr}
\alias{plot.baggr}
\title{Generic plot for baggr package}
\usage{
plot.baggr(x, ...)
}
\arguments{
\item{x}{object of class \code{baggr}}

\item{...}{optional arguments, see \code{baggr_plot}}
}
\value{
ggplot2 object from \code{baggr_plot}
}
\description{
Using generic \code{plot()} on \code{baggr} output invokes \code{\link{baggr_plot}} visual.
See therein for customisation options. Note that plot output is \code{ggplot2} object.`
}
\author{
Witold Wiecek
}
