% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baggr_compare.R
\name{plot.baggr_compare}
\alias{plot.baggr_compare}
\title{Plot method for baggr_compare models}
\usage{
\method{plot}{baggr_compare}(
  x,
  style = "areas",
  arrange = "single",
  interval = 0.95,
  hyper = TRUE,
  transform = NULL,
  order = F,
  ...
)
}
\arguments{
\item{x}{baggr_compare model to plot}

\item{style}{What kind of plot to display (if \code{arrange = "grid"}),
passed to the \code{style} argument in \link{baggr_plot}.}

\item{arrange}{If \code{"single"} (default), generate a single comparison plot;
if \code{"grid"}, display multiple plots side-by-side.}

\item{interval}{probability level used for display of posterior interval}

\item{hyper}{Whether to plot pooled treatment effect
in addition to group treatment effects}

\item{transform}{a function (e.g. exp(), log())
to apply to the values of group (and hyper, if hyper=TRUE)
effects before plotting; when working with effects that are on
log scale, exponent transform is used automatically,
you can plot on log scale by setting transform = identity}

\item{order}{Whether to sort by median treatment effect by group.
If yes, medians from the model with largest range of estimates
are used for sorting.
If not, groups are shown alphabetically.}

\item{...}{ignored for now, may be used in the future}
}
\description{
Allows plots that compare multiple baggr models
that were passed for comparison purposes to baggr compare or
run automatically by baggr_compare
}
