% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baggr_compare.R
\name{plot.baggr_compare}
\alias{plot.baggr_compare}
\title{Plot method for baggr_compare models}
\usage{
\method{plot}{baggr_compare}(x, style = "areas", arrange = "single",
  ...)
}
\arguments{
\item{x}{baggr_compare model to plot}

\item{style}{Style of plot for the baggr_compare model}

\item{...}{ignored for now, may be used in the future}

\item{single}{how to arrange plot display}
}
\description{
Allows plots that compare multiple baggr models
that were passed for comparison purposes to baggr compare or
run automatically by baggr_compar
}
