% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baggr_compare.R
\name{plot.baggr_compare}
\alias{plot.baggr_compare}
\title{Plot method for baggr_compare models}
\usage{
\method{plot}{baggr_compare}(x, style = "areas", arrange = "single",
  interval = 0.95, hyper = T, ...)
}
\arguments{
\item{x}{baggr_compare model to plot}

\item{style}{What kind of plot to display (if \code{arrange = "grid"}),
passed to the \code{style} argument in \link{baggr_plot}.}

\item{arrange}{If \code{"single"} (default), generate a single comparison plot;
if \code{"grid"}, display multiple plots side-by-side.}

\item{interval}{probability level used for display of posterior interval}

\item{hyper}{Whether to plot pooled treatment effect
in addition to group treatment effects}

\item{...}{ignored for now, may be used in the future}
}
\description{
Allows plots that compare multiple baggr models
that were passed for comparison purposes to baggr compare or
run automatically by baggr_compar
}
