% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv.R
\name{plot.baggr_cv}
\alias{plot.baggr_cv}
\title{Plotting method for results of baggr LOO analyses}
\usage{
\method{plot}{baggr_cv}(x, y, ..., add_values = TRUE)
}
\arguments{
\item{x}{output from \link{loocv} that has \code{return_models = TRUE}}

\item{y}{Unused, ignore}

\item{...}{Unused, ignore}

\item{add_values}{logical; if \code{TRUE}, values of \emph{elpd} are printed next to each
study}
}
\value{
\code{ggplot2} plot in similar style to \link{baggr_compare} default plots
}
\description{
Plotting method for results of baggr LOO analyses
}
