% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling_metrics.R
\name{pooling}
\alias{pooling}
\title{Pooling metrics}
\usage{
pooling(bg, metric = "gelman-hill", summary = TRUE)
}
\arguments{
\item{bg}{output of a baggr() function}

\item{metric}{\code{"gelman-hill"} or \code{WIP}}

\item{summary}{logical; if FALSE a whole vector of pooling statistics is returned,
otherwise only the means and intervals}
}
\value{
Matrix with mean and intervals for chosen pooling metric,
each row corresponding to one meta-analysis group.
}
\description{
Compute pooling metrics (of a few different varieties)
given a baggr meta-analysis model
}
\details{
Pooling statistic describes the extent to which group-level estimates of treatment
effect are shrunk toward
average treatment effect in the meta-analysis model.

Different measures of pooling can be estimated and choosing the right one
depends on the research questions.
The default (\code{"gelman-hill"}) "pooling factor" statistic by Gelman & Hill (2007) is
\deqn{\omega(\tau_k) = \frac{WIP}}
}
\references{
Gelman, Andrew, and Jennifer Hill. Data Analysis Using Regression
and Multilevel/Hierarchical Models. 1 edition. Cambridge; New York:
Cambridge University Press, 2006.
}
\author{
Witold Wiecek, Rachael Meager
}
