% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{pp_check.baggr}
\alias{pp_check.baggr}
\title{Posterior predictive checks for baggr model}
\usage{
\method{pp_check}{baggr}(x, type = "dens_overlay", nsamples = 40)
}
\arguments{
\item{x}{Model to check}

\item{type}{type of pp_check. For a list see \link{bayesplot::available_ppc}}

\item{nsamples}{number of samples to compare}
}
\description{
Posterior predictive checks for baggr model
}
