% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.baggr}
\alias{predict.baggr}
\title{Predict method for baggr objects}
\usage{
\method{predict}{baggr}(x, newdata = NULL, allow_new_levels = T,
  nsamples, ...)
}
\arguments{
\item{x}{model fitted via baggr()}

\item{newdata}{new data to use with the fitted model}

\item{allow_new_levels}{whether to allow new groups or levels to be passed to the model.}

\item{...}{other arguments to pass to predict}
}
\description{
Predict method for baggr objects
}
\details{
Predict responses based on the fitted model. Can be used
with the data used to fit the model to perform posterior
predictive checks, or with new data to generate new predictions.
These will include not only the parameter uncertainty but also
the variance that comes as a function of the likelihood. Because
residual error is incorporated, it can be used for visual model
criticism. For linear predictions (e.g. that only incorporate
)
}
