% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.baggr}
\alias{predict.baggr}
\title{Predict method for baggr objects}
\usage{
\method{predict}{baggr}(x, nsamples, newdata = NULL,
  allow_new_levels = T)
}
\arguments{
\item{x}{model to predict from}

\item{nsamples}{Number of samples to draw from the posterior.
Cannot exceed the number of samples in the fitted model.}

\item{newdata}{optional, new data to predict observations from}

\item{allow_new_levels}{whether to allow the model to make predictions
about unobserved groups. Without additional group-level information
the model will use the unconditional, pooled estimate.}

\item{...}{other arguments to pass to predict function}
}
\description{
Predict method for baggr objects
}
