% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.baggr}
\alias{predict.baggr}
\title{Predict treatment effects from baggr model}
\usage{
\method{predict}{baggr}(object, ...)
}
\arguments{
\item{object}{A \code{baggr} class object.}

\item{...}{Additional arguments passed to \link{effect_draw}.}
}
\value{
See \link{effect_draw}.
}
\description{
Alias for \link{effect_draw()}, returning draws from the predictive distribution
for a \code{baggr} model.
}
