% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.baggr}
\alias{predict.baggr}
\title{Predict method for baggr objects}
\usage{
\method{predict}{baggr}(x, newdata = NULL, allow_new_levels = T,
  nsamples, ...)
}
\arguments{
\item{x}{model to predict from}

\item{...}{other arguments to pass to predict}
}
\description{
Predict method for baggr objects
}
\details{
This is currently only implemented for the baggr model
but could also be extended to the others. Currently a WIP, please
report bugs.
}
