% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ma.R
\name{prepare_ma}
\alias{prepare_ma}
\title{Convert from individual to summary data in meta-analyses}
\usage{
prepare_ma(data, standardise = NULL, log = FALSE, cfb = FALSE,
  summarise = TRUE, treatment = "treatment", baseline = NULL,
  group = "group", outcome = "outcome")
}
\arguments{
\item{data}{data.frame of individual-level observations
with columns \code{outcome} (numeric),
\code{treatment} (values 0 and 1) and
\code{group} (numeric, character or factor)}

\item{standardise}{logical; if TRUE, values of outcome
are standardised within each group}
}
\value{
data.frame with columns \code{mu}, \code{se.mu},
\code{tau} and \code{se.tau}
}
\description{
Allows only one-way conversion from full to summary data.
Input must be pre-formatted appropriately.
}
\details{
The conversions are typically not needed and may happen automatically
when data is fed to baggr(). However, this function can be used to explicitly
convert from full to reduced data without analysing it in any model.
}
\seealso{
\code{\link{convert_inputs}}
}
\author{
Witold Wiecek, Rachael Meager
}
