% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ma.R
\name{prepare_ma}
\alias{prepare_ma}
\title{Convert from individual to summary data in meta-analyses}
\usage{
prepare_ma(data, standardise = NULL, log = FALSE, cfb = FALSE,
  summarise = TRUE, treatment = "treatment", baseline = NULL,
  group = "group", outcome = "outcome")
}
\arguments{
\item{data}{data.frame of individual-level observations
with columns \code{outcome} (numeric),
\code{treatment} (values 0 and 1) and
\code{group} (numeric, character or factor)}

\item{standardise}{logical; if TRUE, values of outcome
are standardised within each group}

\item{log}{logical; log-transform data?}

\item{cfb}{logical; calculate change from baseline?}

\item{summarise}{logical; convert to aggregate level data?}

\item{treatment}{name of column with treatment variable}

\item{baseline}{name of column with baseline variable}

\item{group}{name of the column with grouping variable}

\item{outcome}{name of column with outcome variable}
}
\value{
data.frame with columns \code{mu}, \code{se.mu},
\code{tau} and \code{se.tau}
}
\description{
Allows only one-way conversion from full to summary data.
Input must be pre-formatted appropriately.
}
\details{
The conversions are typically not needed and may happen automatically
when data is fed to \code{\link[=baggr]{baggr()}}. However, this function can be used to explicitly
convert from full to reduced data without analysing it in any model.
It can be useful for examining your data.
}
\seealso{
\code{\link[=convert_inputs]{convert_inputs()}} for how data is converted into Stan inputs;
\code{\link[=summarise_quantiles_data]{summarise_quantiles_data()}} for summarising data per quantile
}
\author{
Witold Wiecek
}
