% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_baggr.R
\name{print.baggr}
\alias{print.baggr}
\title{S3 print method for objects of class \code{baggr} (model fits)}
\usage{
\method{print}{baggr}(x, exponent = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{baggr}}

\item{exponent}{if \code{TRUE}, results (for means) are converted to exp scale}

\item{...}{currently unused by this package: further arguments passed
to or from other methods (\code{print}  requirement)}
}
\description{
This \code{print} method for a very concise summary of main model features.
More info is included in the summary of the model and its attributes.
}
