% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_baggr.R
\name{print.baggr}
\alias{print.baggr}
\title{S3 print method for objects of class `baggr` (model fits)}
\usage{
\method{print}{baggr}(x, ...)
}
\arguments{
\item{x}{object of class `baggr`}

\item{...}{currently unused by this pacakge: further arguments passed
to or from other methods (\code{print}  requirement)}
}
\description{
This \code{print} method for a very concise summary of main model features.
More info is included in the summary of the model and its attributes.
}
\details{
TBC
}
