% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv.R
\name{print.baggr_cv}
\alias{print.baggr_cv}
\title{Print baggr cv objects nicely}
\usage{
\method{print}{baggr_cv}(x, digits = 3, ...)
}
\arguments{
\item{x}{baggr_cv object to print}

\item{digits}{number of digits to print}

\item{...}{additional arguments for s3 consistency}
}
\description{
Print baggr cv objects nicely
}
