% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_effects.R
\name{random_effects}
\alias{random_effects}
\title{Extract only random effects from a baggr model}
\usage{
random_effects(...)
}
\arguments{
\item{...}{arguments passed to \link{group_effects}}
}
\description{
This function is a shortcut for \code{group_effects(random_only=TRUE, ...)}
}
