% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{rubin_data}
\alias{rubin_data}
\title{Make model matrix for the rubin data}
\usage{
rubin_data(x, newdata = NULL, allow_new_levels = T)
}
\arguments{
\item{x}{model fitted via baggr()}

\item{newdata}{new data to use with the fitted model}

\item{allow_new_levels}{whether to allow new groups or levels to be passed to the model.}
}
\description{
Make model matrix for the rubin data
}
