% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microcredit.R
\docType{data}
\name{schools}
\alias{schools}
\title{8 schools example}
\format{An object of class \code{data.frame} with 8 rows and 3 columns.}
\usage{
schools
}
\description{
A classic example of aggregate level continuous data in Bayesian hierarchical modelling.
This dataframe contains a column of estimated treatment effects of an SAT prep
program implemented in 8 different schools in the US, and a column of estimated standard errors.
}
\details{
See Gelman et al (1995), Chapter 5, for context and applied example.
}
\references{
Gelman, Andrew, John B. Carlin, Hal S. Stern, and Donald B. Rubin.
Bayesian Data Analysis. Taylor & Francis, 1995.
}
\keyword{datasets}
