% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_dist.R
\name{set_prior_val}
\alias{set_prior_val}
\title{Add prior values to Stan input for baggr}
\usage{
set_prior_val(target, name, prior, p = 1)
}
\arguments{
\item{target}{list object (Stan input) to which prior will be added}

\item{name}{prior name, like \code{hypermean}, \code{hypersd}, \code{hypercor}}

\item{prior}{one of \link{prior} distributions}

\item{p}{number of repeats of the prior, i.e. when P i.i.d. priors are set for
P dimensional parameter as in "mu & tau" type of model}
}
\description{
Add prior values to Stan input for baggr
}
\examples{
set_prior_val(list(), "hypermean", normal(0,1), p = 2)

}
