% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baggr_compare.R
\name{single_comp_plot}
\alias{single_comp_plot}
\title{Plot single comparison plot in baggr_compare style}
\usage{
single_comp_plot(df, title = "", legend = "top", ylab = "", grid = F)
}
\arguments{
\item{df}{data.frame with columns 'group', 'median', 'lci', 'uci', 'model' and
optionally 'parameter'}

\item{title}{'ggtitle'}

\item{legend}{'legend.position'}

\item{ylab}{Y axis label}

\item{grid}{logical; if TRUE, facets by 'parameter' column}
}
\value{
a \code{ggplot2} object
}
\description{
Plot single comparison plot in baggr_compare style
}
