% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_effects.R
\name{study_effects}
\alias{study_effects}
\title{Given a model return treatment effects
(Used as a helper for plotting and printing of results)}
\usage{
study_effects(bg, interval = FALSE)
}
\description{
Given a model return treatment effects
(Used as a helper for plotting and printing of results)
}
