% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_quantiles_data.R
\name{summarise_quantiles_data}
\alias{summarise_quantiles_data}
\title{summarise_quantiles_data}
\usage{
summarise_quantiles_data(data, quantiles, outcome = "outcome",
  group = "group", treatment = "treatment")
}
\arguments{
\item{data}{a data.frame with appropriate grouping, outcome, treatment variables}

\item{quantiles}{a vector of quantiles (between 0 and 1)}

\item{outcome}{character: column name in \code{data} for outcome variable}

\item{group}{character: column name in \code{data} for grouping variable}

\item{treatment}{character: column name in \code{data} for treatment variable}
}
\description{
Given individual level data, return list of summary statistics
Of quantile means and Sigma's, as well as K, N.
Already formatted as a valid input into a Stan model of quantiles
}
\details{
Estimates are obtained externally via \code{\link[quantreg]{rq}}
See \code{\link{baggr}} for documentation of how columns should be formatted.
}
\examples{
summarise_quantiles_data(microcredit_simplified, c(.2, .4, .6),
                         outcome = "consumerdurables")

}
