% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trt_effects.R
\name{treatment_effect}
\alias{treatment_effect}
\alias{hypermean}
\alias{hypersd}
\title{Average treatment effect in a baggr model}
\usage{
treatment_effect(
  bg,
  summary = FALSE,
  transform = NULL,
  interval = 0.95,
  message = TRUE
)

hypermean(
  bg,
  transform = NULL,
  interval = 0.95,
  message = FALSE,
  summary = TRUE
)

hypersd(bg, transform = NULL, interval = 0.95, message = FALSE, summary = TRUE)
}
\arguments{
\item{bg}{a \link{baggr} model}

\item{summary}{logical; if TRUE returns summary statistics as explained below.}

\item{transform}{a transformation to apply to the result, should be an R function;
(this is commonly used when calling \code{treatment_effect} from other
plotting or printing functions)}

\item{interval}{uncertainty interval width (numeric between 0 and 1), if summarising}

\item{message}{logical; use to disable messages prompted by using with
no pooling models}
}
\value{
A list with 2 vectors (corresponding to MCMC samples)
\code{tau} (mean effect) and \code{sigma_tau} (SD). If \code{summary=TRUE},
both vectors are summarised as mean and lower/upper bounds according to
\code{interval}
}
\description{
Average treatment effect in a baggr model

Return the hypermean of the treatments effects

Return the hyper standard deviation of the treatments effects
}
\section{Functions}{
\itemize{
\item \code{hypermean()}: The hypermean of the \code{baggr} model with the interval specified by \code{interval}, with
a default interval of 95\%, as calculated by \link{treatment_effect}.

\item \code{hypersd()}: The hyper-SD of the \code{baggr} model with the interval specified by \code{interval}, with
a default interval of 95\%, as calculated by \link{treatment_effect}.

}}
