% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trt_effects.R
\name{treatment_effect}
\alias{treatment_effect}
\title{Average treatment effect in a baggr model}
\usage{
treatment_effect(bg, transform = NULL)
}
\arguments{
\item{bg}{a \link{baggr} model}

\item{transform}{a transformation to apply to the result, should be an R function;
(this is commonly used when calling \code{group_effects} from other
plotting or printing functions)}
}
\value{
A list with 2 vectors (corresponding to MCMC samples)
\code{tau} (mean effect) and \code{sigma_tau} (SD)
}
\description{
Average treatment effect in a baggr model
}
