% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trt_effects.R
\name{treatment_effect}
\alias{treatment_effect}
\title{Average treatment effect in a baggr model}
\usage{
treatment_effect(bg)
}
\arguments{
\item{bg}{a \link{baggr} model}
}
\value{
A list with 2 vectors (corresponding to MCMC samples)
\code{tau} (mean effect) and \code{sigma_tau} (SD)
}
\description{
Average treatment effect in a baggr model
}
